5 tests crudItem finished

+++ back/app/Http/Controllers/ItemController.php
@@ -67,11 +67,19 @@ class ItemController extends Controller
 
 
         public function destroy($id)
-        {
+        { 
+            if (!auth()->user()->isAdmin) {
+                abort(403, 'Unauthorized action.');
+            }
+        
+            $item = Item::find($id);
+           /*  $item->delete();
+        
+            return redirect()->back(); */ 
             Item::destroy($id);
         
-            return redirect()->route('home');
-        }    
+            return redirect()->route('home'); 
+        }
